Split seat code into its own file
